@@ -11,21 +11,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^12.0.4",
-    "@angular/animations": "^12.0.4",
-    "@angular/cdk": "^12.0.4",
-    "@angular/cli": "^12.2.13",
-    "@angular/common": "^12.0.4",
-    "@angular/compiler": "^12.0.4",
-    "@angular/compiler-cli": "^12.0.4",
-    "@angular/core": "^12.0.4",
+    "@angular-devkit/build-angular": "^13.0.4",
+    "@angular/animations": "^13.0.3",
+    "@angular/cdk": "^13.0.3",
+    "@angular/cli": "^13.0.4",
+    "@angular/common": "^13.0.3",
+    "@angular/compiler": "^13.0.3",
+    "@angular/compiler-cli": "^13.0.3",
+    "@angular/core": "^13.0.3",
     "@angular/flex-layout": "^9.0.0-beta.31",
-    "@angular/forms": "^12.0.4",
-    "@angular/material": "^12.0.4",
-    "@angular/material-moment-adapter": "^12.0.4",
-    "@angular/platform-browser": "^12.0.4",
-    "@angular/platform-browser-dynamic": "^12.0.4",
-    "@angular/router": "^12.0.4",
+    "@angular/forms": "^13.0.3",
+    "@angular/material": "^13.0.3",
+    "@angular/material-moment-adapter": "^13.0.3",
+    "@angular/platform-browser": "^13.0.3",
+    "@angular/platform-browser-dynamic": "^13.0.3",
+    "@angular/router": "^13.0.3",
     "@babel/core": "^7.14.0",
     "@fortawesome/angular-fontawesome": "^0.8.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
@@ -33,6 +33,9 @@
     "@ngrx/effects": "^10.1.2",
     "@ngrx/store": "^10.1.2",
     "@ngrx/store-devtools": "^10.1.2",
+    "@turf/boolean-point-in-polygon": "^6.5.0",
+    "@turf/intersect": "^6.5.0",
+    "@turf/line-intersect": "^6.5.0",
     "@types/d3": "^5.16.4",
     "@types/ol": "^6.5.3",
     "@types/ua-parser-js": "^0.7.36",
@@ -54,6 +57,7 @@
     "file-saver": "^2.0.5",
     "flex-layout-sass": "^1.0.8",
     "glob": "^7.1.6",
+    "install": "^0.13.0",
     "intl": "^1.2.5",
     "jwt-decode": "^2.2.0",
     "moment": "^2.29.1",
@@ -63,6 +67,7 @@
     "ngx-operators": "^2.1.0",
     "ngx-toastr": "^13.2.1",
     "nouislider": "^15.5.0",
+    "npm": "^8.3.0",
     "ol": "^6.5.0",
     "popper.js": "^1.16.1",
     "postcss": "^8.3.4",
@@ -76,7 +81,7 @@
     "timers": "^0.1.1",
     "tippy.js": "^5.2.1",
     "tslib": "^2.2.0",
-    "typescript": "^4.2.4",
+    "typescript": "^4.4.4",
     "ua-parser-js": "^0.7.28",
     "uuid": "^3.4.0",
     "wnumb": "^1.2.0",
@@ -84,7 +89,7 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular/language-service": "^12.0.4",
+    "@angular/language-service": "^13.0.3",
     "@ngrx/schematics": "^10.1.2",
     "@types/jasmine": "^3.6.11",
     "@types/jasminewd2": "^2.0.8",
