@@ -77,6 +77,7 @@ export class AppComponent implements OnInit, OnDestroy, AfterViewInit {
     private dialog: MatDialog,
     private notificationService: services.NotificationService,
     private sarviewsService: services.SarviewsEventsService,
+    private mapService: services.MapService
   ) {}
 
   public ngAfterViewInit(): void {
@@ -445,8 +446,8 @@ export class AppComponent implements OnInit, OnDestroy, AfterViewInit {
   public onClearSearch(): void {
     this.store$.dispatch(new scenesStore.ClearScenes());
     this.store$.dispatch(new scenesStore.SetSelectedSarviewsEvent(''));
+    this.mapService.clearDrawLayer();
     this.store$.dispatch(new uiStore.CloseResultsMenu());
-
     this.searchService.clear(this.searchType);
   }
 
