@@ -13,7 +13,7 @@ export const RtcGammaJobType: Hyp3JobType = {
   productTypes: [{
     dataset: sentinel_1,
     productTypes: [
-      'SLC', 'GRD_HD', 'GRD_HS'
+      'SLC', 'GRD_HD'
     ],
     beamModes: ['IW'],
     polarizations: [
@@ -82,25 +82,25 @@ export const RtcGammaJobType: Hyp3JobType = {
     default: false,
     info: `Apply an Enhanced Lee speckle filter.`
   }, {
-    name: 'Include DEM',
+    name: 'DEM',
     apiName: 'include_dem',
     type: JobOptionType.CHECKBOX,
     default: false,
     info: `Include the DEM file in the product package.`
   }, {
-    name: 'Include Inc. Angle Maps',
+    name: 'Incidence Angle Maps',
     apiName: 'include_inc_map',
     type: JobOptionType.CHECKBOX,
     default: false,
     info: `Include the incidence angle maps (local and ellipsoidal) in the product package.`
   }, {
-    name: 'Include Scattering Area',
+    name: 'Scattering Area',
     apiName: 'include_scattering_area',
     type: JobOptionType.CHECKBOX,
     default: false,
     info: `Include the scattering area in the product package.`
   }, {
-    name: 'Include RGB',
+    name: 'RGB Decomposition',
     apiName: 'include_rgb',
     type: JobOptionType.CHECKBOX,
     default: false,
@@ -146,7 +146,7 @@ export const InsarGammaJobType: Hyp3JobType = {
     default: '20x4',
     info: `Number of looks to take in range and azimuth.`
   }, {
-    name: 'Apply Water Mask',
+    name: 'Water Mask',
     apiName: 'apply_water_mask',
     type: JobOptionType.TOGGLE,
     default: false,
@@ -155,7 +155,7 @@ export const InsarGammaJobType: Hyp3JobType = {
       for phase unwrapping.
     `
   }, {
-    name: 'Include DEM',
+    name: 'DEM',
     apiName: 'include_dem',
     type: JobOptionType.CHECKBOX,
     default: false,
@@ -163,7 +163,7 @@ export const InsarGammaJobType: Hyp3JobType = {
       Include the DEM file in the product package.
     `
   }, {
-    name: 'Include Inc. Angle Maps',
+    name: 'Incidence Angle Maps',
     apiName: 'include_inc_map',
     type: JobOptionType.CHECKBOX,
     default: false,
@@ -171,7 +171,7 @@ export const InsarGammaJobType: Hyp3JobType = {
       Include the incidence angle maps (local and ellipsoidal) in the product package.
     `
   }, {
-    name: 'Include Look Vectors',
+    name: 'Look Vectors',
     apiName: 'include_look_vectors',
     type: JobOptionType.CHECKBOX,
     default: false,
@@ -179,7 +179,7 @@ export const InsarGammaJobType: Hyp3JobType = {
       Include the look vector theta and phi files in the product package.
     `
   }, {
-    name: 'Include Displacement Maps',
+    name: 'Displacement Maps',
     apiName: 'include_displacement_maps',
     type: JobOptionType.CHECKBOX,
     default: false,
@@ -187,7 +187,7 @@ export const InsarGammaJobType: Hyp3JobType = {
       Include displacement maps (line-of-sight and vertical) in the product package.
     `
   }, {
-    name: 'Include Wrapped Phase',
+    name: 'Wrapped Phase',
     apiName: 'include_wrapped_phase',
     type: JobOptionType.CHECKBOX,
     default: false,
@@ -224,7 +224,7 @@ export const AutoRift: Hyp3JobType = {
   productTypes: [{
     dataset: sentinel_1,
     productTypes: [
-      'SLC', 'GRD_HD', 'GRD_HS'
+      'SLC', 'GRD_HD'
     ],
     beamModes: ['IW'],
     polarizations: [
