@@ -164,7 +164,12 @@ export class Hyp3Service {
 
       hyp3ableProducts.forEach(product => {
         const prodType = product[0].metadata.productType;
-        byProdType[prodType].push(product);
+        byProdType[prodType].push(product?.sort((a, b) => {
+          if (a.metadata.date < b.metadata.date) {
+            return -1;
+          }
+          return 1;
+        }));
       });
 
       const byProductType: models.Hyp3ableByProductType[] = Object.entries(byProdType).map(([productType, prods]) => ({
