@@ -17,6 +17,7 @@ export enum MapActionType {
   SET_BROWSE_OVERLAY_OPACITY = '[Map] Set Browse Overlay Opacity',
   CLEAR_BROWSE_OVERLAYS = '[Map] Clear All Browse Overlays On Map',
   DRAW_NEW_POLYGON = '[Map] Set has user drawn new polygon',
+  TOGGLE_OVERVIEW_MAP = '[Map] Toggle Overview Map'
 }
 
 export class SetStreetView implements Action {
@@ -81,6 +82,12 @@ export class DrawNewPolygon implements Action {
   public readonly type = MapActionType.DRAW_NEW_POLYGON;
 }
 
+export class ToggleOverviewMap implements Action {
+  public readonly type = MapActionType.TOGGLE_OVERVIEW_MAP;
+
+  constructor(public payload: boolean) {}
+}
+
 
 export type MapActions =
   | SetMapView
@@ -94,4 +101,5 @@ export type MapActions =
   | SetBrowseOverlays
   | SetBrowseOverlayOpacity
   | ClearBrowseOverlays
-  | DrawNewPolygon;
+  | DrawNewPolygon
+  | ToggleOverviewMap;
